pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "metapulsar-analysis"
version = "0.1.0"
description = "Multi-PTA pulsar timing data combination framework for gravitational wave detection"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Rutger van Haasteren", email = "rutger@vhaasteren.com"},
    {name = "Wang-Wei Yu", email = "wangwei.yu@aei.mpg.de"},
]
keywords = ["pulsar", "timing", "gravitational-waves", "pta", "astronomy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "numpy>=1.20.0",
    "astropy>=5.0.0",
    "scipy>=1.7.0",
    "pint-pulsar>=0.9.0",
    "enterprise-pulsar>=3.0.0",
    "h5pulsar @ git+https://github.com/vhaasteren/h5pulsar@main",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=22.0",
    "mypy>=0.900",
    "flake8>=4.0",
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
]
libstempo = [
    "libstempo>=2.0.0",
]
analysis = [
    "xarray>=0.20.0",
    "arviz>=0.12.0",
    "la-forge>=0.1.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
]

[project.urls]
Homepage = "https://gitlab.aei.uni-hannover.de/vhaasteren/metapulsar-analysis"
Repository = "https://gitlab.aei.uni-hannover.de/vhaasteren/metapulsar-analysis"
Documentation = "https://metapulsar-analysis.readthedocs.io"
Issues = "https://gitlab.aei.uni-hannover.de/vhaasteren/metapulsar-analysis/-/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
